split data into partial classes
