+++ Crossword.Data/GetPuzzleDataAsync.cs
@@ -7,7 +7,7 @@ namespace Crossword.Data;
 /// <summary>
 /// Calls the Crossword.API to get the Puzzledata as a string
 /// </summary>
-public static class GetPuzzleDataAsync
+public static partial class GetPuzzleDataAsync
 {
     #region CallDataApiAsync
 
@@ -60,39 +60,6 @@ public static class GetPuzzleDataAsync
         {
             logger.LogError(ex, ex.Message);
             return string.Empty;
-        }
-
-    }
-
-    #endregion
-
-
-    #region GetCrosswordDataAsync
-
-    /// <summary>
-    /// GetCrosswordDataAsync
-    /// </summary>
-    /// <returns></returns>
-    public static async Task<string?> GetCrosswordDataAsync()
-    {
-        //Init the logger and get the active config
-        using var logger = new SerilogLogger(ConfigurationHelper.ActiveConfiguration);
-
-        //Call the API to get the puzzledata....otherwise use default values
-        try
-        {
-            logger.LogInformation("Start GetCrosswordDataAsync()");
-
-            //call the API
-            var apiResponse = await CallDataApiAsync();
-
-            //check what was returned
-            return string.IsNullOrEmpty(apiResponse) ? GameConstants.DefaultPuzzleData : apiResponse;
-        }
-        catch (Exception ex)
-        {
-            logger.LogError(ex, ex.Message);
-            throw;
         }
 
     }
